Add support for chunked transfer encoding
Support for chunked transfer encoding is mandated by many web servers.
For the protocol details, see

  https://en.wikipedia.org/wiki/Chunked_transfer_encoding

and RFC7203, section 4.1:

  https://tools.ietf.org/html/rfc7230#section-4.1

Support for this transfer scheme is now built in to
HttpClient::responseBody(), and the state machine handles
Content-Length transmissions as well. That has the benefit that
timing delays are handled fine for both cases.

Note that this replaces #7.

Fixes #3