Write body in chunks of 1024
There seems to be a limit in how many bytes Client::write() may take
at once, and data is not sent if it is exceeded. I'm not entriely
sure where that limit comes from, but I have the suspicion is has to
do with the MTU.

A value of 1024 works fine in my test. If that turns out to not cover
all cases, the threshold can be made configurable later on.